fix: Move setup and update workers
