+++ playwright.config.ts
@@ -17,8 +17,7 @@ export default defineConfig({
   forbidOnly: !!process.env.CI,
   /* Retry on CI only */
   retries: process.env.CI ? 2 : 0,
-  /* Opt out of parallel tests on CI. */
-  workers: process.env.CI ? 1 : undefined,
+  workers: 4,
   /* Reporter to use. See https://playwright.dev/docs/test-reporters */
   reporter: process.env.CI ? "github" : "list",
   /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
@@ -54,19 +53,12 @@ export default defineConfig({
   },
 
   projects: [
-    // Our global setup to configure the Nextcloud docker container
-    {
-      name: "setup",
-      testMatch: /setup\.ts$/,
-    },
     {
       name: "chromium",
       testMatch: /.*\.spec\.ts/,
-      testIgnore: /setup\.ts$/,
       use: {
         ...devices["Desktop Chrome"],
       },
-      dependencies: ["setup"],
     },
   ],
 });
